extract base eval types
